realworld-freesio-scala: Article Service In Progress

+++ realworld-starter-kit/realworld-freesio-scala/src/main/scala/com/realworld/articles/persistence/ArticlesRepository.scala
@@ -6,12 +6,13 @@ import freestyle.tagless.tagless
 
 @tagless(true)
 trait ArticlesRepository[F[_]] {
+
   import ArticleEntity._
 
   def insertArticle(articleEntity: ArticleEntity, user_id: Long, tags: List[String]): F[Option[Long]]
   def getArticle(slug: String, user_id: Long): F[Option[ArticleResponse]]
-  def getRecentFollowedArticles(limit: Int, offset: Int, user_id: Long) : F[List[ArticleResponse]]
-  def getRecentArticles(user_id: Long,limit: Int, offset: Int): F[List[ArticleResponse]]
+  def getRecentFollowedArticles(limit: Long, offset: Long, user_id: Long) : F[List[ArticleResponse]]
+  def getRecentArticles(user_id: Long,limit: Long, offset: Long): F[List[ArticleResponse]]
   def updateArticle(articleEntity: ArticleEntity): F[Int]
   def deleteArticle(slug: String, user_id: Long): F[Int]
   def init: F[Int]
